requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Optional but recommended
umap-learn>=0.5.3
shap>=0.42.0
tabulate>=0.9.0

# For Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
